Fix (un)signed variable warnings (lib)

+++ src/get_focused_window_linux.cc
@@ -63,7 +63,7 @@ static char *get_property(Display *disp, Window win, Atom xa_prop_type,
     }
 
     // null terminate the result to make string handling easier
-    tmp_size = (ret_format / (32 / sizeof(long))) * ret_nitems; // NOLINT
+    tmp_size = (static_cast<unsigned>(ret_format) / (32 / sizeof(long))) * ret_nitems;
     ret = reinterpret_cast<char *>(malloc(tmp_size + 1));
     if (!ret) {
         XFree(ret_prop);
